Fix the issue that cannot input more than 4 characters. It is due to uint16_t issue, changes all uint16, uint24 to uint32

+++ include/internal/chewingutil.h
@@ -37,7 +37,7 @@ int SymbolChoice(ChewingData *pgdata, int sel_i);
 int HaninSymbolInput(ChewingData *pgdata);
 void WriteChiSymbolToCommitBuf(ChewingData *pgdata, ChewingOutput *pgo, int len);
 int ReleaseChiSymbolBuf(ChewingData *pgdata, ChewingOutput *);
-int AddChi(uint16_t phone, uint16_t phoneAlt, ChewingData *pgdata);
+int AddChi(uint32_t phone, uint32_t phoneAlt, ChewingData *pgdata);
 int CallPhrasing(ChewingData *pgdata, int all_phrasing);
 int MakeOutputWithRtn(ChewingOutput *pgo, ChewingData *pgdata, int keystrokeRtn);
 int MakeOutput(ChewingOutput *pgo, ChewingData *pgdata);
